requirements.txt
# requirements.txt
#
# This file lists the Python packages required to run your Google Meet Recording Processor.
# It ensures that all necessary dependencies are installed in your environment.

# Streamlit for the graphical user interface (--ui mode)
streamlit>=1.45.0

# Google API Client Libraries (required for "Streaming from Google Meet" mode)
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0

# Other libraries
git+https://github.com/openai/whisper.git
opencv-python
easyocr
tqdm
requests
torch
